Refactor AWSSync pipeline (#67)
* Changed comments

* Refactored create move

* Refractored create_account

* Changed test

* Added my functions

* Add api talker function for describe_create_account_status

* Deleted create_account function

* Added test

* Fix describe_create_account_status docstring

* Changed describe function

* Finished basic test

* Create structure of refactored pipeline

* Finished refactor create move

* Black formattting

* Duplicate code

* Reorganize files and complete merge

* Refactor according to finished create_and_move_account, aws data extraction

* first attempt at printing an error box when something fails in the pipeline (functional but not too pretty).

* Hotfixes pipeline for working sprint 3 demo

* Add pipeline test cases

* Fix 100% code coverage

* Fix tests for error box in pipeline

* Added the current policy to the pipeline

* Add API talker function and test for untag_resource AWS API call

* Remove most debug messages when testing

* Rename awssync_refactored to awssync

* Increase sleep time between requesting account info

* Implement pull request feedback

---------
